# backend/swagger.yaml
openapi: 3.0.0
info:
    description: API documentation of Wealth Tracker API.
    version: 1.0.0-oas3
    title: Wealth Tracker API Documentation
    contact:
        email: gradimbuyi@outlook.com
    license:
        name: MIT
        url: https://opensource.org/license/mit
...
paths:
...
    /accounts:
        get:
            tags:
                - users
            summary: Get all the accounts associated with logged in user.
            operationId: getAccounts
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Returned accounts associated with User.
                '500':
                    description: Internal Server Error.
    /account:
        post:
            tags:
                - users
            summary: Create account for logged in User.
            operationId: createAccount
            security:
                - bearerAuth: []
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/createAccount'
                description: Required Fields for account creation.
            responses:
                '200':
                    description: Returned account associated with User.
                '500':
                    description: Internal server error.
    /account/{id}:
        get:
            tags:
                - users
            summary: Get specific account associated with logged in User.
            operationId: getAccount
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  description: The id of the account to retrieve.
                  schema:
                      type: string
            responses:
                '200':
                    description: Returned account associated with User.
                '500':
                    description: Internal server error.
        put:
            tags:
                - users
            summary: Update specific account associated with logged in User.
            operationId: updateAccount
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  description: The id of the account to update.
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/updateAccount'
            responses:
                '200':
                    description: Updated account associated with User.
                '400':
                    description: Failed to update account. Error in request body format.
                '500':
                    description: Internal server error.
        delete:
            tags:
                - users
            summary: Delete specific account associated with logged in User.
            operationId: deleteAccount
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  description: The id of the account to delete.
                  schema:
                      type: string
            responses:
                '200':
                    description: Deleted account associated with User.
                '400':
                    description: Failed to delete account. Error in request body format.
                '500':
                    description: Internal server error.
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        updateAccount:
            type: object
            properties:
                accountName:
                    type: string
                    example: Wells Fargo Checking
                accountType:
                    type: string
                    example: Checking
                accountNumber:
                    type: number
                    example: 20762971795
                accountInstitution:
                    type: string
                    example: Wells Fargo
                balance:
                    type: number
                    example: 2062.15
                currency:
                    type: string
                    example: USD
                isActive:
                    type: boolean
                    example: true
        createAccount:
            type: object
            required:
                - accountName
                - accountType
                - accountNumber
                - accountInstitution
                - balance
            properties:
                accountName:
                    type: string
                    example: Wells Fargo Checking
                accountType:
                    type: string
                    example: Checking
                accountNumber:
                    type: number
                    example: 20762971795
                accountInstitution:
                    type: string
                    example: Wells Fargo
                balance:
                    type: number
                    example: 2062.15
                currency:
                    type: string
                    example: USD
                isActive:
                    type: boolean
                    example: true
...
servers:
    - url: http://wealthtrackit.tech/api
      description: Production server
    - url: https://virtserver.swaggerhub.com/grady-4ce/WealthTracker/1.0.0
      description: Mock server (for testing)